+++ lib/standard/facets/random.rb
@@ -5,28 +5,13 @@ require 'facets/kernel/maybe'
 # = Randomization Extensions
 #
 # This library extends Array, String, Hash and other classes with randomization
-# methods. Most of the methods are of one of two kinds. Either they "pick" a
-# random element from the reciever or they randomly "shuffle" the reciever.
-#
-# The most common example is Array#shuffle, which simply randmomizes the
-# order of an array's elements.
-#
-#   [1,2,3].shuffle  #~> [2,3,1]
-#
-# The other methods do similar things for their respective classes.
-#
-# The classes are all extended via mixins which have been created within
-# Ruby's Random class.
+# methods.
 #
 # Credit for this work is due to:
 #
 # * Ilmari Heikkinen
 # * Christian Neukirchen
 # * Thomas Sawyer
-#
-# NOTE: Inline QED tests are almost pointless here b/c... yea, it's random.
-# So indtead of the usual `#=>` we use `#~>` which means approx. equal and
-# prevens QED from making a hard assertion.
 
 class Random
 
@@ -36,10 +21,6 @@ class Random
 
     public :number
   end
-
-  ##def self.number(*args)
-  ##  ::Kernel.rand(*args)
-  ##end
 
   # Module method to generate a random letter.
   #
@@ -68,36 +49,19 @@ class Random
     # CREDIT: Lavir the Whiolet, Thomas Sawyer
     def at_rand
       first, last = first(), last()
-      if first.respond_to?(:random_delta)  # TODO: work on this!!!
+      if first.respond_to?(:random_delta)
         begin
           first.random_delta(last, exclude_end?)
         rescue
-          to_a.at_rand
+          to_a.sample
         end
       else
-        to_a.at_rand
+        to_a.sample
       end
-      ##elsif first.respond_to?(:succ)
-      ##  # optimized algorithm
-      ##  if (Fixnum === first || Bignum === first) &&
-      ##     (Fixnum === last  || Bignum === last)
-      ##    last -= 1 if exclude_end?
-      ##    return nil if last < first
-      ##    return Random.number(last - first + 1) + first
-      ##  end
-      ##  # standard algorithm
-      ##  return to_a.at_rand
-      ##elsif Numeric === first && Numeric === last
-      ##  return nil if last < first
-      ##  return nil if exclude_end? && last == first
-      ##  return (last - first) * Random.number + first
-      ##else
-      ##  return nil
-      ##end
     end
   end
 
-  # Random extensions fo Integer class.
+  # Random extensions for Integer class.
   #
   module IntegerExtensions
     #
@@ -124,40 +88,19 @@ class Random
   # Random extensions for Array class.
   #
   module ArrayExtensions
-    # Return a random element from the array.
-    #
-    #   [1, 2, 3, 4].at_rand           #~> 2
-    #   [1, 2, 3, 4].at_rand           #~> 4
-    #
-    def at_rand
-      at(Random.number(size))
-    end
 
-    # Same as #at_rand, but acts in place removing a
-    # random element from the array.
+    # Destructive version of Array#sample. Removes and returns
+    # a random element, or n random exclusive elements.
     #
     #   a = [1,2,3,4]
-    #   a.at_rand!       #~> 2
+    #   a.sample!        #~> 2
     #   a                #~> [1,3,4]
     #
-    def at_rand!
-      return delete_at( Random.number( size ) )
-    end
-
-    # Similar to #at_rand, but will return an array of randomly
-    # picked exclusive elements if given a number.
-    def pick(n=nil)
-      if n
-        a = self.dup
-        a.pick!(n)
-      else
-        at(Random.number(size))
-      end
-    end
-
-    # Similar to #at_rand!, but given a number will return
-    # an array of exclusive elements.
-    def pick!(n=nil)
+    #   a = [1,2,3,4,5]
+    #   a.sample!(3)     #~> [2,4,1]
+    #   a                #~> [3,5]
+    #
+    def sample!(n=nil)
       if n
         if n > self.size
           r = self.dup
@@ -173,6 +116,30 @@ class Random
       end
     end
 
+    # @deprecated Use Array#sample instead.
+    def at_rand
+      warn "Array#at_rand is deprecated. Use Array#sample instead.", uplevel: 1
+      sample
+    end
+
+    # @deprecated Use Array#sample! instead.
+    def at_rand!
+      warn "Array#at_rand! is deprecated. Use Array#sample! instead.", uplevel: 1
+      sample!
+    end
+
+    # @deprecated Use Array#sample instead.
+    def pick(n=nil)
+      warn "Array#pick is deprecated. Use Array#sample instead.", uplevel: 1
+      n ? sample(n) : sample
+    end
+
+    # @deprecated Use Array#sample! instead.
+    def pick!(n=nil)
+      warn "Array#pick! is deprecated. Use Array#sample! instead.", uplevel: 1
+      sample!(n)
+    end
+
     # Random index.
     #
     def rand_index
@@ -184,14 +151,6 @@ class Random
     # elements, otherwise returns a random number of elements
     # upto the size of the Array.
     #
-    # By defualt the returned values are exclusive of
-    # each other, but if _exclusive_ is set to <tt>false</tt>,
-    # the same values can be choosen more than once.
-    #
-    # When _exclusive_ is <tt>true</tt> (the default) and the
-    # _number_ given is greater than the size of the array,
-    # then all values are returned.
-    #
     #   [1, 2, 3, 4].rand_subset(1)        #~> [2]
     #   [1, 2, 3, 4].rand_subset(4)        #~> [2, 1, 3, 4]
     #   [1, 2, 3, 4].rand_subset           #~> [1, 3, 4]
@@ -200,7 +159,6 @@ class Random
     def rand_subset(number=nil, exclusive=true)
       number = Random.number(size) unless number
       number = number.to_int
-      #return self.dup if (number >= size and exlusive)
       return sort_by{rand}.slice(0,number) if exclusive
       ri =[]; number.times { |n| ri << Random.number(size) }
       return values_at(*ri)
@@ -223,36 +181,6 @@ class Random
       end
     end
 
-    # Randomize the order of an array.
-    #
-    #   [1,2,3,4].shuffle  #~> [2,4,1,3]
-    #
-    def shuffle
-      dup.shuffle!
-      #sort_by{Random.number}
-    end
-
-    # As with #shuffle but modifies the array in place.
-    # The algorithm used here is known as a Fisher-Yates shuffle.
-    #
-    #   a = [1,2,3,4]
-    #   a.shuffle!
-    #
-    #   a  #~> [2,4,1,3]
-    #
-    # CREDIT Niel Spring
-    def shuffle!
-      s = size
-      each_index do |j|
-        i = Random.number(s-j)
-        #self[j], self[j+i] = self[j+i], self[j]
-        tmp = self[j]
-        self[j] = self[j+i]
-        self[j+i] = tmp
-      end
-      self
-    end
-
   end
 
   # Random extensions for Hash class.
@@ -260,7 +188,7 @@ class Random
   module HashExtensions
     # Returns a random key.
     #
-    #   {:one => 1, :two => 2, :three => 3}.pick_key  #~> :three
+    #   {:one => 1, :two => 2, :three => 3}.rand_key  #~> :three
     #
     def rand_key
       keys.at(Random.number(keys.size))
@@ -278,11 +206,9 @@ class Random
       return k
     end
 
-    alias_method( :pick_key, :rand_key! )
-
     # Returns a random key-value pair.
     #
-    #   {:one => 1, :two => 2, :three => 3}.pick  #~> [:one, 1]
+    #   {:one => 1, :two => 2, :three => 3}.rand_pair  #~> [:one, 1]
     #
     def rand_pair
       k = rand_key
@@ -301,8 +227,6 @@ class Random
       return k,v
     end
 
-    alias_method( :pick_pair, :rand_pair! )
-
     # Returns a random hash value.
     #
     #   {:one => 1, :two => 2, :three => 3}.rand_value  #~> 2
@@ -315,19 +239,14 @@ class Random
     # Deletes a random key-value pair and returns the value.
     #
     #   a = {:one => 1, :two => 2, :three => 3}
-    #   a.at_rand!  #~> 2
-    #   a           #~> {:one => 1, :three => 3}
+    #   a.rand_value!  #~> 2
+    #   a              #~> {:one => 1, :three => 3}
     #
     def rand_value!
       k,v = rand_pair
       delete( k )
       return v
     end
-
-    alias_method( :pick, :rand_value! )
-
-    alias_method( :at_rand, :rand_value )
-    alias_method( :at_rand!, :rand_value! )
 
     # Returns a copy of the hash with _values_ arranged
     # in new random order.
@@ -364,15 +283,12 @@ class Random
     # Class-level methods.
     module Self
       # Returns a randomly generated string. One possible use is
-      # password initialization. Takes a max legnth of characters
+      # password initialization. Takes a max length of characters
       # (default 8) and an optional valid char Regexp (default /\w\d/).
       #
       #   String.random    #~> 'dd4qed4r'
       #
       # CREDIT George Moschovitis
-      #--
-      # TODO: This is not very efficient. Better way?
-      #++
       def random(max_length = 8, char_re = /[\w\d]/)
         raise ArgumentError.new('second argument must be a regular expression') unless char_re.is_a?(Regexp)
         string = ""
@@ -387,18 +303,17 @@ class Random
       #
       # CREDIT: Guido De Rosa
       def random_binary(n_bytes)
-        ( Array.new(n_bytes){ rand(0x100) } ).pack('c*') 
+        ( Array.new(n_bytes){ rand(0x100) } ).pack('c*')
       end
     end
 
     # Return a random separation of the string.
-    # Default separation is by charaacter.
+    # Default separation is by character.
     #
     #   "Ruby rules".at_rand(' ')  #~> ["Ruby"]
     #
     def at_rand( separator=// )
-      #separator = self.class.patterns( separator )
-      self.split(separator,-1).at_rand
+      self.split(separator,-1).sample
     end
 
     # Return a random separation while removing it
@@ -409,7 +324,6 @@ class Random
     #   s                  #~> "rules"
     #
     def at_rand!( separator=// )
-      #separator = self.class.patterns( separator )
       a = self.shatter( separator )
       w = []; a.each_with_index { |s,i| i % 2 == 0 ? w << s : w.last << s }
       i = Random.number(w.size)
@@ -447,8 +361,8 @@ class Random
       Random.number(size)
     end
 
-    # Return the string with seperated sections arranged
-    # in a random order. The default seperation is by character.
+    # Return the string with separated sections arranged
+    # in a random order. The default separation is by character.
     #
     #   "Ruby rules".shuffle  #~> "e lybRsuur"
     #